% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conshaplotypes.R
\name{conshaplotypes}
\alias{conshaplotypes}
\title{conshaplotypes}
\usage{
conshaplotypes(
  samples,
  primers,
  thr,
  min.seq.len,
  min.rd,
  fileTable,
  poolTable
)
}
\arguments{
\item{thr}{Minimum threshold of abundance to enter in multiple alignment.}

\item{min.seq.len}{Minimum length of sequences to enter in intersection.}

\item{min.rd}{Minimum nº of reads required in fasta files.}

\item{fileTable}{Data frame ...}

\item{poolTable}{Data frame ...}
}
\description{
fitxer rawconshaplos
}
\examples{
min.seq.len <- 150
min.rd <-   1
thr <- 0.2

}
