% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexMID.R
\name{demultiplexMID}
\alias{demultiplexMID}
\title{Split reads by MID sequence}
\usage{
demultiplexMID(
  flashffiles,
  samples,
  mids,
  maxdif = 1,
  mid.start = 1,
  mid.end = 40
)
}
\arguments{
\item{flashffiles}{Vector including the names of FLASH filtered files, with fastq extension.}

\item{samples}{Data frame with relevant information about the samples of the sequencing experiment, including
\code{Patient.ID, MID, Primer.ID, Region, RefSeq.ID}, and \code{Pool.Nm} columns.}

\item{mids}{Data frame with the association between MID identifiers and their sequences.}

\item{maxdif}{Number of mismatches allowed between MID and read sequences.}

\item{mid.start}{Expected start position for MID in sequence.}

\item{mid.end}{Expected end position for MID in sequence.}
}
\value{
A list containing the following:
  \item{nreads}{A table with the number of reads
    identified for each MID.}
  \item{by.pools}{A table with the coverage of reads demultiplexed by pool.}
  After execution, a FASTA file for each combination of MID and pool will be saved in the splits folder,
  including its associated reads. Additionaly, two report files will be generated:
  \enumerate{
  \item \code{SplidByMIDs.barplots.pdf}: Includes a first barplot representing \code{nreads} data values,
    and a second plot with the \code{by.pools} data values.
  \item \code{SplidByMIDs.Rprt.txt}: Includes the same data tables returned by the function.}
}
\description{
Demultiplex reads by identifying MID sequences within windows of expected positions in the sequenced reads.
  MIDs are 10 base-length oligonucleotides that allow the identification of samples from different patients
  or origins.

It is important to note that MID sequences will be not trimmed from reads, they are only identified for
  association with each sample.
}
\details{
demultiplexMID
}
\note{
Files indicated in \code{flashffiles} must be located in a directory named flashFiltDir, and
  a directory for resulting FASTA files with demultiplexed reads must be named as splitDir.
  Also, a reports folder must be created in the project environment, whose path will be
  named as repDir.
}
\examples{
repDir <- "./reports"
splitDir <- "./splits"
flashFiltDir <- "./flashFilt"
flashffiles <- list.files(flashFiltDir)
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                      colClasses="character",stringsAsFactors=F)
mids <- read.table("./data/mids.csv", sep="\t", header=T,
                   stringsAsFactors=F)
maxdif <- 1
mid.start <- 1
mid.end <- 40
dem.res<-demultiplexMID(flashffiles,samples,mids,maxdif,mid.start,mid.end)
}
\seealso{
\code{\link{FiltbyQ30}}
}
\author{
Alicia Aranda
}
