% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiltbyQ30.R
\name{FiltbyQ30}
\alias{FiltbyQ30}
\title{FiltbyQ30}
\usage{
FiltbyQ30(max.pct = 0.05, flashfiles, flashres)
}
\arguments{
\item{max.pct}{The maximum percentage of bases below Q30 allowed in reads (by default,5\%).}

\item{flashfiles}{Vector including the names of files that are going to be processed,
with fastq extension.}

\item{flashres}{Table of results obtained after the execution of \code{\link{R1R2toFLASH}}
function.}
}
\value{
A \code{\link{data.frame}} object containing FLASH and Filtering results,
  but also two report files:
  \enumerate{
  \item \code{FiltQ30.barplot.pdf}: Includes a first Bar plot representing raw reads,
    extended reads by FLASH and filtered reads, and a second Bar plot with
    the yield by process for each pool.
  \item \code{FiltQ30_report.txt}: The same data returned by the function.
}
  The results table obtained includes two new columns with respect to FLASH results
    table, named FiltQ30 (number of filtered reads) and Raw (total sequencing reads).
}
\description{
This function applies \code{\link{FastqStreamer}} over a fastq file and removes
  all reads that have a defined fraction of bases below Q30.
}
\details{
Files indicated in \code{flashfiles} must be located in a folder named flashDir.
  Also, a reports folder must be created in the project environment, whose path will be
  named as repDir.
}
\note{
This function is designed to be applied after \code{\link{R1R2toFLASH}} function from
  the same package. If \code{flashres} is not specified but FLASH extension was previously
  done, the function will try to load the FLASH results table from the reports folder.
}
\examples{
runDir <- "./run"
runfiles <- list.files(runDir)
flash <- "C:/FLASH/flash.exe"
flashDir <- "./flash"
flashfiles <- list.files(flashDir)
flashres <- R1R2toFLASH(runfiles,flash)
filtres <- FiltbyQ30(max.pct=0.05,flashfiles,flashres)

}
\seealso{
\code{\link{R1R2toFLASH}}, \code{\link{FastqStreamer}}
}
