% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapplyhack.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{Execution of parallel:mclapply() on Windows machines}
\usage{
mclapply.hack(..., mc.cores = NULL)
}
\arguments{
\item{mc.cores}{Number of cores to use for parallelization.}
}
\description{
Mimics forking on Windows machines just like it is done with
  \code{\link{mclapply}} in Mac or Linux.
}
\note{
The function code was extracted from \code{post-10-mclapply-hack.R} written by Nathan VanHoudnos, and was only
  added in this package for parallelization using Windows operating system. See the original code
  in <https://github.com/nathanvan/mcmc-in-irt/blob/master/post-10-mclapply-hack.R>
}
\seealso{
\code{\link{mclapply}}
}
\author{
Nathan VanHoudnos
}
