% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCbyPosplot.R
\name{plot.QC}
\alias{plot.QC}
\title{plot.QC}
\usage{
\method{plot}{QC}(fvnm1, fvnm2, snm, SW = FALSE, FL = FALSE)
}
\arguments{
\item{fvnm1, fvnm2}{Matrix or array containing Phred score values by read position for
5, 25, 50, 75 and 95% quantiles.}

\item{snm}{Character indicating the name of the pool of evaluated reads.}

\item{SW}{Logical indicating if the plot should include quality profile by SW (Sliding Window).}

\item{FL}{Logical indicating if the first argument corresponds to FLASH extended reads scores.}
}
\description{
This function generates QC plots by read position
}
