% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInDels.R
\name{PlotInDels}
\alias{PlotInDels}
\title{Check for insertions and deletions in samples}
\usage{
PlotInDels(machfiles, pm.res)
}
\arguments{
\item{machfiles}{Vector including the paths of files generated by \code{\link{ConsHaplotypes}} function,
with fna extension.}

\item{pm.res}{The list returned by \code{\link{demultiplexPrimer}}, including \code{fileTable}
and \code{poolTable} data frames.}
}
\value{
After execution, a file named \code{GapsBarPlots.pdf} will be saved in the reports folder,
  including the plots generated for all samples. In each plot, insertions are represented by red lines
  and deletions are represented by blue lines.
}
\description{
Generates insertion/deletion plots for each evaluated sample
  from consensus haplotypes' sequences.
}
\note{
This function is designed to be applied at the end of the quality assessment analysis and requires
  the previous execution of \code{\link{demultiplexPrimer}} and \code{\link{ConsHaplotypes}} functions
  from the same package.
}
\examples{
## Execute demultiplexing by primer
splitDir <- "./splits"
splitfiles <- list.files(splitDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
pm.res <- demultiplexPrimer(splitfiles,samples,primers)

## Obtain consensus haplotypes (default parameters)
trimDir <- "./trim"
trimfiles <- list.files(trimDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
int.res <- ConsHaplotypes(trimfiles, pm.res, thr, min.seq.len)

## Apply function
mach.Dir <- "./MACH"
machfiles <- list.files(mach.Dir,recursive=TRUE, full.names=TRUE, include.dirs=TRUE)
PlotInDels(machfiles,pm.res)
}
