% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolQC.R
\name{PoolQC}
\alias{PoolQC}
\title{PoolQC}
\usage{
PoolQC(runfiles, flashfiles, samples, primers)
}
\arguments{
\item{runfiles}{Vector including the names of Illumina MiSeq Raw Data files, often with fastq.gz extension}

\item{samples}{Data frame}

\item{primers}{Data frame}

\item{flahsfiles}{Vector including the names of FLASH processed files, with fastq extension}
}
\description{
PoolQC
}
\details{
Files indicated in \code{runfiles} and \code{flashfiles} arguments must be located in different folders,
  named runDir and flashDir respectively.
}
