% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCReadLenPeaks.R
\name{PoolQCbyRead}
\alias{PoolQCbyRead}
\title{PoolQCbyRead}
\usage{
PoolQCbyRead(flashfiles, samples, primers)
}
\arguments{
\item{samples}{Data frame}

\item{primers}{Data frame}

\item{flahsfiles}{Vector including the names of FLASH processed files, with fastq extension}
}
\description{
PoolQCbyRead
}
\details{
Files indicated in \code{flashfiles} must be located in a folder named flashDir.
  Also, a reports folder must be created in the project environment, whose path will be
  named as repDir.
}
