% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCbyPosplot.R
\name{QCplot}
\alias{QCplot}
\title{QCplot}
\usage{
QCplot(fvnm1, fvnm2, snm, SW = FALSE, FL = FALSE)
}
\arguments{
\item{fvnm1, fvnm2}{Matrix or array containing Phred score values by read position for
0.05, 0.25, 0.5, 0.75 and 0.95 quantiles.}

\item{snm}{Character indicating the name of the pool of evaluated reads.}

\item{SW}{Logical indicating if the plot should include quality profile by SW (Sliding Window).}

\item{FL}{Logical indicating if the first argument corresponds to FLASH extended reads scores.}
}
\value{
plot QC
}
\description{
This function generates QC plots by read position
}
