% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCscores.R
\name{QCscores}
\alias{QCscores}
\title{QCscores}
\usage{
QCscores(flnm, ln = 301, byPos = FALSE, byRead = FALSE)
}
\arguments{
\item{flnm}{Fastq file to be evaluated.}

\item{ln}{Amplicon length (by default 301 nt).}

\item{byPos}{Logical indicating if QC by position should be done.}

\item{byRead}{Logical indicating if QC by read should be done.}
}
\value{
If argument \code{byPos} is TRUE, the function returns a list including
  the following parameters:
  \enumerate {
  \item \code{fvnq}: }
}
\description{
This function applies \code{\link{FastqStreamer}} over a fastq file and returns quality
  control mesures (Phred scores), either by position or by read.
}
