% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conshaplotypes.R
\name{ConsHaplotypes}
\alias{ConsHaplotypes}
\title{Generate consensus haplotypes}
\usage{
ConsHaplotypes(trimfiles, pm.res, thr = 0.2, min.seq.len = 150)
}
\arguments{
\item{trimfiles}{Vector including the paths of demultiplexed files by specific primer, with fna extension.}

\item{pm.res}{The list returned by \code{\link{demultiplexPrimer}}, including \code{fileTable}
and \code{poolTable} data frames.}

\item{thr}{Threshold to filter haplotypes at minimum abundance before multiple alignment.}

\item{min.seq.len}{Threshold to filter haplotypes at minimum length before intersection.}
}
\value{

}
\description{
Computes the intersection of forward and reverse strand haplotypes and generates some report files.
}
\details{
This function is designed to be used after the execution of \code{\link{demultiplexPrimer}} function
  from the same package. After the generation of FASTA files containing forward and reverse strand reads
  for the evaluated samples, \code{\link{ConsHaplotypes}} executes multiple alignment with \code{\link{muscle}} and
  returns the consensus haplotypes using \code{\link{IntersectStrandHpls}}, that will be saved using
  the helper function \code{\link{SaveHaplotypes}}.
}
\examples{
splitDir <- "./splits"
# Save the file names with complete path
splitfiles <- list.files(splitDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
# Get data
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                      colClasses="character",stringsAsFactors=F)
mids <- read.table("./data/mids.csv", sep="\t", header=T,
                   stringsAsFactors=F)
# Apply previous function from QA analysis
pm.res <- demultiplexPrimer(splitfiles,samples,primers)
# Save the files generated by previous function
trimDir <- "./trim"
trimfiles <- list.files(trimDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
# Define necessary parameters
min.seq.len <- 150
min.rd <-   1
thr <- 0.2
int.res <- ConsHaplotypes(trimfiles, pm.res, thr, min.seq.len)

}
\seealso{
\code{\link{muscle}},\code{\link{IntersectStrandHpls}},\code{\link{demultiplexPrimer}},\code{\link{SaveHaplotypes}}
}
\author{
Alicia Aranda
}
