% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conshaplotypes.R
\name{ConsHaplotypes}
\alias{ConsHaplotypes}
\title{ConsHaplotypes}
\usage{
ConsHaplotypes(pm.res, thr, min.seq.len)
}
\arguments{
\item{pm.res}{The list returned by \code{\link{demultiplexprimer}}, including \code{fileTable}
and \code{poolTable} data frames.}

\item{thr}{Threshold to filter haplotypes at minimum abundance before intersection.}

\item{min.seq.len}{Threshold to filter haplotypes at minimum length before intersection.}
}
\value{
The function returns a \code{\link{data.frame}} object containing the intersection results
  for each combination of patient and pool, including the initial number of reads, filtered out reads
  (for being below a given frequency threshold or unique to a single strand), overlapping frequency
  between both strands and the common reads (in percentage and nº of reads).

  After execution, two fasta files for each combination of sample and pool will be saved in the MACH folder;
  the first includes multiple alignment between forward and reverse strand haplotypes, and the second
  includes the forward and reverse strands intersected. Additionaly, some report files will be generated:
  \enumerate{
  \item \code{MA.Intersects-SummRprt.txt}: Includes the sumary results by reads number after abundance filter and
    strand intersection.
  \item \code{MA.Intersects.plots.pdf}: Includes different barplots for each sample representing the frequency of
    forward, reverse and intersected strand haplotypes.
  \item \code{IntersectBarplots.pdf}: Includes different barplots for all combinations of patient and pool,
    representing the number of intersected and filtered out reads, the intersection yield and global yield.}
}
\description{
Archivos almacenados en la carpeta global del proyecto: “muscle.log”. Es el archivo donde se guardan los
parámetros de la ejecución de muscle.
Archivos almacenados en la carpeta reports: "IntersectBarplots.pdf", "IntersectedReads.RData",
"MA.Intersects.plots.pdf", "MA.Intersects-SummRprt.txt".
Archivos almacenados en la carpeta tmp: “muscleInFile.fna”, “muscleOutFile.fna”. Se generan al ejecutar el
programa muscle, son los archivos de entrada y salida del programa.
Archivos almacenados en la carpeta MACH: 4 archivos .fna por paciente (2 por pool) →
los que empiezan por MAfwrv corresponden a los resultados del alineamiento múltiple (muscle) de los haplotipos
de cada muestra (forward y reverse) después de realizar el filtro por abundancia de 0.2%; los que empiezan por
MACHpl02 se generan después de hacer la intersección entre cadenas FW y RV, incluyen las secuencias fasta de los
haplotipos que han coincidido entre las cadenas.
}
\details{
Computes the intersection of forward and reverse strand haplotypes and generates some report files.

This function is designed to be used after the execution of \code{\link{demultiplexprimer}} function
  from the same package. After the generation of fasta files containing forward and reverse strand reads
  for the evaluated samples, \code{\link{ConsHaplotypes}} executes multiple alignment with \code{\link{muscle}} and
  returns the consensus haplotypes with \code{\link{IntersectStrandHpls}}.
}
\examples{
# Apply previous function from QA analysis
pm.res <- demultiplexprimer(splitfiles,samples,primers)
# Define necessary parameters
min.seq.len <- 150
min.rd <-   1
thr <- 0.2
ConsHaplotypes(pm.res, thr, min.seq.len)

}
\seealso{
\code{\link{muscle}}
}
