% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conshaplotypes.R
\name{ConsHaplotypes}
\alias{ConsHaplotypes}
\title{Generate consensus haplotypes}
\usage{
ConsHaplotypes(trimfiles, pm.res, thr = 0.2, min.seq.len = 150)
}
\arguments{
\item{trimfiles}{Vector including the names of demultiplexed files by specific primer, with fna extension.}

\item{pm.res}{The list returned by \code{\link{demultiplexPrimer}}, including \code{fileTable}
and \code{poolTable} data frames.}

\item{thr}{Threshold to filter haplotypes at minimum abundance before multiple alignment.}

\item{min.seq.len}{Threshold to filter haplotypes at minimum length before intersection.}
}
\value{
The function returns a \code{\link{data.frame}} object containing the intersection results
  for each combination of patient and pool, including the initial number of reads, filtered out reads
  (for being below a given frequency threshold or unique to a single strand), overlapping frequency
  between both strands and the common reads (in percentage and nº of reads).

  After execution, two FASTA files for each combination of sample and pool will be saved in the MACH folder;
  the first includes multiple alignment between forward and reverse strand haplotypes, and the second
  includes the forward and reverse strands intersected. Additionaly, some report files will be generated:
  \enumerate{
  \item \code{MA.Intersects-SummRprt.txt}: Includes the sumary results by reads number after abundance filter and
    strand intersection.
  \item \code{MA.Intersects.plots.pdf}: Includes different barplots for each sample representing the frequency of
    forward, reverse and intersected strand haplotypes.
  \item \code{IntersectBarplots.pdf}: Includes different barplots for all combinations of patient and pool,
    representing the number of intersected and filtered out reads, the intersection yield and global yield.}
}
\description{
Computes the intersection of forward and reverse strand haplotypes and generates some report files.
}
\details{
ConsHaplotypes

This function is designed to be used after the execution of \code{\link{demultiplexPrimer}} function
  from the same package. After the generation of FASTA files containing forward and reverse strand reads
  for the evaluated samples, \code{\link{ConsHaplotypes}} executes multiple alignment with \code{\link{muscle}} and
  returns the consensus haplotypes using \code{\link{IntersectStrandHpls}}, that will be saved using
  the helper function \code{\link{SaveHaplotypes}}.
}
\note{
Files indicated in \code{trimfiles} must be located in a directory named trimDir, and
  a directory for resulting FASTA files with intersected reads and the results of multiple alignment
  must be named as mach.Dir. Also, a reports folder must be created in the project environment, whose path will be
  named as repDir.
}
\examples{
splitDir <- "./splits"
splitfiles <- list.files(splitDir)
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                      colClasses="character",stringsAsFactors=F)
mids <- read.table("./data/mids.csv", sep="\t", header=T,
                   stringsAsFactors=F)
# Apply previous function from QA analysis
pm.res <- demultiplexPrimer(splitfiles,samples,primers)
# Define necessary parameters
min.seq.len <- 150
min.rd <-   1
thr <- 0.2
int.res <- ConsHaplotypes(pm.res, thr, min.seq.len)

}
\seealso{
\code{\link{muscle}},\code{\link{IntersectStrandHpls}},\code{\link{demultiplexPrimer}},\code{\link{SaveHaplotypes}}
}
\author{
Alicia Aranda
}
