% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeFLASH.R
\name{executeFLASH}
\alias{executeFLASH}
\title{executeFLASH}
\usage{
executeFLASH(R1, R2, outfile = "flash.fastq", chunck.sz = 1e+06)
}
\arguments{
\item{R1}{Path for R1 reads file.}

\item{R2}{Path for R2 reads file.}

\item{outfile}{File path for FLASH output. If it is not specified, the fastq file
generated will be saved in current working directory.}

\item{chunck.sz}{Chunck size to be used by \code{\link{FastqStreamer}} function}
}
\value{
This function returns a matrix containing de number of reads extended
  and not extended by FLASH. Additionaly, a fastq file with extended reads
  will be saved to outfile path.
}
\description{
Runs FLASH to extend R1 and R2 reads.
This function is only defined for correct execution of \code{\link{R1R2toFLASH}} function from the same package.
}
\seealso{
\code{\link{R1R2toFLASH}}
}
