% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeFLASH.R
\name{executeFLASH}
\alias{executeFLASH}
\title{Run FLASH to extend R1 and R2 reads.}
\usage{
executeFLASH(R1, R2, flash, flash.opts, outfile = "./flash.fastq")
}
\arguments{
\item{R1}{Path for R1 reads file.}

\item{R2}{Path for R2 reads file.}

\item{flash}{Folder path containing FLASH executable.}

\item{flash.opts}{Character indicating FLASH options that will be part of the execution command.}

\item{outfile}{File path for FLASH output. If it is not specified, the fastq file
generated will be saved in current working directory.}
}
\value{
This function returns a matrix containing de number of reads extended
  and not extended by FLASH. Additionaly, a fastq file with extended reads
  will be saved to outfile path. Further FLASH output files will be saved in a new folder named "tmp".
}
\description{
With R1 and R2 files generated by paired-end sequencing, the function executes FLASH program
  to obtain the number of extended and not-extended reads.
}
\note{
This function is defined for correct execution of \code{\link{R1R2toFLASH}} function from
  the same package, where all arguments are defined automatically.
}
\seealso{
\code{\link{R1R2toFLASH}}
}
\author{
Alicia Aranda
}
