% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolQC.R
\name{PoolQC}
\alias{PoolQC}
\title{PoolQC}
\usage{
PoolQC(runfiles, flashfiles, samples, primers)
}
\arguments{
\item{runfiles}{Character indicating the Raw Data files from Illumina MiSeq, often with fastq.gz extension}

\item{samples}{Data frame}

\item{primers}{Data frame}

\item{flahsfiles}{Character indicating the files processed by FLASH program, with fastq extension}
}
\description{
PoolQC
}
