% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolQCbyPos.R
\name{PoolQCbyPos}
\alias{PoolQCbyPos}
\title{Evaluate QC by position

This function evaluates fastq files before and after the execution of FLASH program to extend
  paired-end reads, and returns QC by position plots in pdf format.
It can be applied also after filtering FLASH fastq files by Phred Score.}
\usage{
PoolQCbyPos(flashfiles, samples, primers, runfiles)
}
\arguments{
\item{flashfiles}{Vector including the names of FLASH processed/filtered files, with fastq extension.}

\item{samples}{Data frame with relevant information about the samples of the sequencing experiment, including
\code{Patient.ID, MID, Primer.ID, Region, RefSeq.ID}, and \code{Pool.Nm} columns.}

\item{primers}{Data frame with information about the \emph{primers} used in the experiment, including
\code{Ampl.Nm, Region, Primer.FW, Primer.RV, FW.pos, RV.pos, FW.tpos, RV.tpos, Aa.ipos},
  and \code{Aa.lpos} columns.}

\item{runfiles}{Vector including the names of Illumina MiSeq Raw Data files, often with fastq.gz extension.
If the function is applied for filtered fastq files, this argument must be NA or missing.}
}
\value{
After execution, a pdf file for each pool used in the experiment will be saved in the
  reports folder, and a message indicating that the files are generated will appear in console.
}
\description{
PoolQCbyPos
}
\note{
Files indicated in \code{runfiles} and \code{flashfiles} arguments must be located in different folders,
  whose paths will be named as runDir and flashDir respectively. Also, a reports folder must be created
  in the project environment, whose path will be named as repDir.
}
\examples{
runDir <- "./run"
flashDir <- "./flash"
repDir <- "./reports"
runfiles <- list.files(runDir)
flashfiles <- list.files(flashDir)
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                     colClasses="character",stringsAsFactors=F)
primers <- read.table("./data/mids.csv", sep="\t", header=T,
                      stringsAsFactors=F)
PoolQCbyPos(flashfiles,samples,primers,runfiles)
}
\author{
Alicia Aranda
}
