% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolQCbyRead.R
\name{PoolQCbyRead}
\alias{PoolQCbyRead}
\title{Evaluate QC by read}
\usage{
PoolQCbyRead(flashfiles, samples, primers)
}
\arguments{
\item{flashfiles}{Vector including the paths of FLASH processed files, with fastq extension.}

\item{samples}{Data frame with relevant information about the samples of the sequencing experiment, including
\code{Patient.ID, MID, Primer.ID, Region, RefSeq.ID}, and \code{Pool.Nm} columns.}

\item{primers}{Data frame with information about the \emph{primers} used in the experiment, including
\code{Ampl.Nm, Region, Primer.FW, Primer.RV, FW.pos, RV.pos, FW.tpos, RV.tpos, Aa.ipos},
  and \code{Aa.lpos} columns.}
}
\value{
After execution, 3 pdf files will be saved in the reports folder; the first 2 include
  the QC by read plots for each pool used in the experiment, and the 3rd shows the read length
  distribution for each pool. A message indicating that the files are generated will appear in console.
}
\description{
This function evaluates fastq files after the execution of FLASH program to extend
  paired-end reads, and returns QC by read plots in pdf format.
}
\examples{
flashDir <- "./flash"
repDir <- "./reports"
# Save the file names with complete path
flashfiles <- list.files(flashDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
# Get data
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                     colClasses="character",stringsAsFactors=F)
primers <- read.table("./data/primers.csv", sep="\t", header=T,
                      stringsAsFactors=F)
PoolQCbyRead(flashfiles,samples,primers)
}
\seealso{
\code{\link{R1R2toFLASH}},\code{\link{QCscores}}
}
\author{
Alicia Aranda
}
