% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolQCbyRead.R
\name{PoolQCbyRead}
\alias{PoolQCbyRead}
\title{Evaluate QC by read}
\usage{
PoolQCbyRead(flashfiles, samples, primers, ncores = 1)
}
\arguments{
\item{flashfiles}{Vector including the paths of FLASH processed files, with fastq extension.}

\item{samples}{Data frame with relevant information to identify the samples of the sequencing experiment, including
\code{Patient.ID, MID, Primer.ID, Region, RefSeq.ID}, and \code{Pool.Nm} columns.}

\item{primers}{Data frame with information about the \emph{primers} used in the experiment, including
\code{Ampl.Nm, Region, Primer.FW, Primer.RV, FW.pos, RV.pos, FW.tpos, RV.tpos, Aa.ipos},
  and \code{Aa.lpos} columns.}

\item{ncores}{Number of cores to use for parallelization with \code{\link{mclapply.hack}}.}
}
\value{
After execution a message will appear in console, indicating that the following
report files have been generated (and saved in a reports folder):
\enumerate{
  \item{\code{PoolQCbyRead_PoolName.pdf}: This file is generated for each pool used in the
    experiment, after extracting its name from \code{samples} data frame. The pdf includes
    includes a representation of bases below Q30 (in nº of reads and percentage) by read.}
  \item{\code{PoolReadLengths.pdf}: Includes one plot for each pool representing the read length distribution.}}
}
\description{
This function evaluates fastq files after the execution of the FLASH program to extend
  paired-end reads, and returns QC by read plots in pdf format. The results of this function are
  important for defining the maximum fraction of bases below Q30 allowed in reads, which will
  be used in \code{\link{FiltbyQ30}} function.
}
\examples{
flashDir <- "./flash"
repDir <- "./reports"
# Save the file names with complete path
flashfiles <- list.files(flashDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
# Get data
samples <- read.table("./data/samples.csv", sep="\t", header=T,
                     colClasses="character",stringsAsFactors=F)
primers <- read.table("./data/primers.csv", sep="\t", header=T,
                      stringsAsFactors=F)
PoolQCbyRead(flashfiles,samples,primers)
}
\seealso{
\code{\link{R1R2toFLASH}}, \code{\link{QCscores}}
}
\author{
Alicia Aranda
}
