% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplot.R
\name{QCplot}
\alias{QCplot}
\title{QC plot by read position}
\usage{
QCplot(fvnm1, fvnm2, snm, SW = FALSE, FL = FALSE)
}
\arguments{
\item{fvnm1, fvnm2}{Matrix or array containing Phred score values by read position for
0.05, 0.25, 0.5, 0.75 and 0.95 quantiles.
If a QC by position plot for R1 and R2 files is required, both arguments are needed, one
for each file.}

\item{snm}{Character indicating the name of the pool of evaluated reads. Only required when
both \code{fvnm1} and \code{fvnm2} are provided.}

\item{SW}{Logical indicating if the plot should include quality profile by SW (Sliding Window).
If \code{TRUE}, Phred Scores are computed as moving averages for windows of 10 base length
slid along the sequence.}

\item{FL}{Logical indicating if the first argument corresponds to the scores of FLASH extended reads.
If \code{TRUE}, only \code{fvnm1} argument is required.}
}
\value{
A scatter plot with desired values will be shown in the active plots window.
}
\description{
Draws a scatter plot with Phred score values by read position.
}
\details{
QCplot

\code{fvnm1} and \code{fvnm2} arguments are obtanied from the \code{\link{QCscores}}
  function with argument \code{byPos=TRUE}.
}
\note{
This function is only defined for correct execution of \code{\link{PoolQCbyPos}}
 function from the same package.
}
\examples{
flashDir <- "./flash"
flashfiles <- list.files(flashDir)
lst1 <- QCscores(file.path(flashDir,flashfiles[1]),byPos=T)
fvnm1 <- lst1$fvnq
QCplot(fvnm1,FL=TRUE) # QC plot for FLASH fastq file
QCplot(fvnm1,SW=TRUE,FL=TRUE) # QC plot by SW for FLASH fastq file
}
\seealso{
\code{\link{PoolQCbyPos}}, \code{\link{QCscores}}
}
\author{
Alicia Aranda
}
