% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1R2toFLASH.R
\name{R1R2toFLASH}
\alias{R1R2toFLASH}
\title{Run FLASH to extend paired-end reads and generate report graphs.}
\usage{
R1R2toFLASH(runfiles, flash, min.ov = 20, max.ov = 300, err.lv = 0.1)
}
\arguments{
\item{runfiles}{Vector including the paths of Illumina MiSeq Raw Data files, often with fastq.gz extension.}

\item{flash}{File path of FLASH executable.}

\item{min.ov}{Minimum overlap (in nt) between R1 and R2 reads.}

\item{max.ov}{Maximum overlap (in nt) between R1 and R2 reads.}

\item{err.lv}{Mismatch fraction accepted in overlapping.}
}
\value{
The function returns a \code{\link{data.frame}} object containing FLASH results
  for sequenced regions.

After the execution, a fastq file with extended reads for each pool will be saved in
  a new folder named flash. Additionaly, two files will be saved in a reports folder:
  \enumerate{
  \item{\code{FLASH_barplot.pdf}: Bar plots representing extended vs not extended reads
    and the yield of the process for each pool.}
  \item{\code{FLASH_report.txt}: Includes the data returned by the function with used FLASH parameters.}
}
}
\description{
This function applies \code{\link{executeFLASH}} over R1 and R2 reads
  for multiple sample pools and returns a file with the extended reads for each pool.
}
\examples{
runDir <- "./run"
flash <- "./FLASH/flash.exe"
# Save the file names with complete path
runfiles<-list.files(runDir,recursive=TRUE,full.names=TRUE,include.dirs=TRUE)
min.ov <- 20
max.ov <- 300
err.lv <- 0.1
flashres <- R1R2toFLASH(runfiles,flash,min.ov,max.ov,err.lv)
}
\seealso{
\code{\link{executeFLASH}}
}
\author{
Alicia Aranda
}
