% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1R2toFLASH.R
\name{R1R2toFLASH}
\alias{R1R2toFLASH}
\title{R1R2toFLASH}
\usage{
R1R2toFLASH(
  runfiles,
  flash,
  min.len = 200,
  min.ov = 20,
  max.ov = 300,
  err.lv = 0.1,
  chunck.sz = 1e+06
)
}
\arguments{
\item{runfiles}{Character indicating which files are going to be processed, often with fastq.gz extension}

\item{flash}{Folder path containing FLASH executable}

\item{min.len}{Minimum length to consider a sequence}

\item{min.ov}{Minimum overlap (in nt) between R1 and R2}

\item{max.ov}{Maximum overlap (in nt) between R1 and R2}

\item{err.lv}{Mismatch fraction accepted in overlapping}

\item{chunck.sz}{Chunck size to be used by \code{\link{FastqStreamer}} function}
}
\value{
The function returns a \code{\link{data.frame}} object containing FLASH results
  for your sequenced regions, but also two report files:
  \enumerate{
  \item \code{FLASH_barplot.pdf}: Bar plots representing extended vs not extended reads
    and the yield of the process for each pool.
  \item \code{FLASH_report.txt}: Includes the data returned by the function with FLASH parameters used.
}
}
\description{
A function that runs FLASH program to extend paired-end reads and generates some report graphics.
}
\details{
Files indicated in \code{runfiles} must be located in a folder named runDir.
  Also, a reports folder must be created in the project environment, whose path will be
  named as repDir.
}
\examples{
runDir <- "C:/run"
flash <- "C:/FLASH/flash.exe"
runfiles <- list.files(runDir)
R1R2toFLASH(runfiles,flash,min.len=200,min.ov=20,max.ov=300,err.lv=0.10,chunck.sz=1.e6)
}
